Split demo apps into separate modules
Fixes #62
