+++ demo/README.md
@@ -1,9 +1,15 @@
 # CloudBridge Demo
 
+The demo is split into platform modules that share `:demo:common`:
+
+- `:demo:android` — Android app
+- `:demo:desktop` — Desktop (JVM) app
+- `:demo:web` — Web (JS/Wasm) app
+
 ### Desktop
 
-`gradlew :demo:run`
+`gradlew :demo:desktop:run`
 
 ### Web
 
-`gradlew :demo:jsBrowserDevelopmentRun`
+`gradlew :demo:web:jsBrowserDevelopmentRun`
