+++ demo/android/src/main/AndroidManifest.xml
