+++ demo/shared/build.gradle.kts
@@ -1,0 +1,47 @@
+@file:Suppress("OPT_IN_USAGE")
+
+plugins {
+    alias(libs.plugins.kotlin.multiplatform)
+    alias(libs.plugins.android.kotlin.multiplatform.library)
+    alias(libs.plugins.compose.compiler)
+    alias(libs.plugins.compose.multiplatform)
+}
+
+kotlin {
+    android {
+        namespace = "nl.jacobras.cloudbridge.demo.shared"
+        compileSdk = libs.versions.android.compileSdk.get().toInt()
+        minSdk = libs.versions.android.minSdk.get().toInt()
+    }
+    jvm("desktop")
+    js { browser() }
+    wasmJs { browser() }
+
+    sourceSets {
+        commonMain.dependencies {
+            api(projects.library)
+            api(libs.androidx.viewmodel)
+            api(libs.compose.foundation)
+            api(libs.compose.material3)
+            api(libs.compose.runtime)
+            api(libs.compose.ui)
+            api(libs.kotlin.coroutines.core)
+            implementation(libs.compose.icons.extended)
+            implementation(libs.compose.resources)
+            implementation(libs.compose.ui.tooling.preview)
+            implementation(libs.humanReadable)
+            implementation(libs.kermit)
+            implementation(libs.multiplatform.settings)
+        }
+        webMain.dependencies {
+            implementation(libs.kotlinx.browser)
+        }
+        androidMain.dependencies {
+            implementation(libs.multiplatform.settings.no.arg)
+        }
+    }
+
+    compilerOptions {
+        optIn.add("kotlin.time.ExperimentalTime")
+    }
+}
