+++ demo/web/webpack.config.d/watch.js
