+++ gradle.properties
@@ -2,9 +2,4 @@ org.gradle.configuration-cache=true
 kotlin.daemon.jvmargs=-Xmx3g
 org.gradle.jvmargs=-Xmx3g -Dfile.encoding=UTF-8
 
-# The demo applies the classic 'com.android.application' plugin alongside Kotlin Multiplatform.
-# Since AGP 9.0 that combination requires letting KMP manage Kotlin and using the legacy Android DSL.
-android.builtInKotlin=false
-android.newDsl=false
-
 version=0.5.0
